requirements.txt
apispec==6.3.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
flask-smorest==0.42.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
numpy==1.25.0
packaging==23.1
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
tzdata==2023.3
webargs==8.2.0
Werkzeug==2.3.6
